requirements.txt
matplotlib==3.8.2
matplotlib-inline==0.1.3
numpy==1.26.2
pandas==2.0.0
pyyaml_env_tag==0.1
seaborn==0.13.1
